refactor(ui): move SettingsScreen into settings/ subpackage
Prep step for breaking the 1062-LOC SettingsScreen god-file into
section composables. Moving the orchestrator first means each
extracted section can be a sibling internal-visible file in
hivens.ui.screens.settings without import gymnastics.

Pure rename + import-site fix in AppLayout. Behavior unchanged.
