+++ client-ui/src/desktopMain/kotlin/hivens/ui/screens/settings/SettingsScreen.kt
@@ -1,4 +1,4 @@
-package hivens.ui.screens
+package hivens.ui.screens.settings
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
